Make NonEmptyPassword accessible in other storage implementations
